backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<1.5.0
alembic>=1.7.0,<1.8.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.5,<0.1.0
pydantic>=1.8.0,<2.0.0
asyncpg<0.28.0,>=0.27.0
redis>=4.5.0,<4.6.0
fastapi-limiter>=0.1.5,<0.2.0
fastapi-cache2>=0.2.1,<0.3.0
itsdangerous>=2.1.2,<3.0.0
tenacity>=8.2.0,<9.0.0